Consolidate vector setter benchmarks
Move the generic SetChunkValue benchmark helper and type matrix into one
file so the exact BIGINT case does not duplicate setup.

Cover []byte-to-VARCHAR and string-to-BLOB fallback writes explicitly.
Canonical fast paths preserve their allocation behavior, while supported
cross-type fallback writes may box. The benchmark suite now exposes that
boundary instead of implying every accepted conversion is allocation-free.
